Change getScore INT to FLOAT + create Mobility Heuristic

+++ LOA_lab3/src/Algo/Heuristic/HeuristicInterface.java
@@ -3,5 +3,5 @@ package Algo.Heuristic;
 import util.GameState;
 
 public interface HeuristicInterface {
-	abstract public int getScore(GameState gameState);
+	abstract public float getScore(GameState gameState);
 }
